hbgd_data_store_server/requirements.txt
beautifulsoup4==4.5.1
bokeh==0.12.3
colorcet==0.9.1
crispy-forms-foundation==0.5.4
Django==1.10.5
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-docs==0.2.1
django-extensions==1.7.5
django-tables2==1.2.6
factory_boy==2.8.1
pandas==0.18.1
psycopg2==2.6.2
pytest==3.0.5
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-selenium==1.9.0
selenium==3.8.1
sphinx==1.5.3
sphinx-rtd-theme==0.2.0
toolz==0.8.2
pytest-flake8==0.8.1
python-coveralls==2.9.0
whitenoise==3.2.2
